+++ resources/api.rb
@@ -13,6 +13,7 @@ attribute :host, kind_of: String, default: lazy { node['nexus3']['api']['host'] 
 attribute :endpoint, kind_of: String, default: lazy { node['nexus3']['api']['endpoint'] }
 attribute :ignore_failure, kind_of: [TrueClass, FalseClass], default: lazy { node['nexus3']['api']['ignore_failure'] }
 attribute :live_stream, kind_of: [TrueClass, FalseClass], default: lazy { node['nexus3']['api']['live_stream'] }
+attribute :wait, kind_of: Integer, default: lazy { node['nexus3']['api']['wait'] }
 attribute :sensitive, kind_of: [TrueClass, FalseClass] # , default: true - see initialize below
 
 # Chef will override sensitive back to its global value, so set default to true in init