+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0 2016-08-11
+
+- Wait for Nexus Rest API endpoint to respond
+- Replace vmoptions attribute with template
+
 ## 0.3.0 2016-08-08
 
 - Fix #4 Support Windows install